% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tracks.R
\name{add_matrix_track}
\alias{add_matrix_track}
\title{Add column tracks as a matrix to a \code{ggheatmap}}
\usage{
add_matrix_track(
  gghm,
  track_columns,
  track_colors = "Blues",
  colors_title = "value",
  rows_title = NULL,
  track_prop = 0.3,
  fontsize = 11,
  track_pos = "bottom",
  legend_action = "collect",
  colorbar_dir = "vertical"
)
}
\arguments{
\item{gghm}{An ggplot object of class \code{ggheatmap}.}

\item{track_columns}{A vector with names of columns to be plotted.}

\item{track_colors}{A named list where names are the same as \code{track_columns}.
Can be either:
\itemize{
\item A valid palette used by \code{RColorBrewer}. See: \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::display.brewer.all()}}
\item A named vector of colors passed to \code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual()}}.
}}

\item{colors_title}{A title for the color legend}

\item{rows_title}{A title for the variables in the rows}

\item{track_prop}{A number between 0 and 1, representing the height
proportion between new tracks and the heatmap.}

\item{fontsize}{Base fontsize for plot, which will be used by the theme.
Ultimately passed to \code{\link[ggplot2:ggtheme]{ggplot2::theme_minimal()}} as \code{base_size}.}

\item{track_pos}{One of: 'bottom' or 'top'.}

\item{legend_action}{A string specifying how guides should be treated in the layout.
See: guides in \code{\link[patchwork:plot_layout]{patchwork::plot_layout()}}.}

\item{colorbar_dir}{one of "vertical" or "horizontal". See \link[ggplot2:guide_colourbar]{ggplot2::guide_colorbar}}
}
\description{
add_matrix_track() is a shortcut for building a new panel of \code{ggplot2::geom_tile()}s
with \code{track_columns} and adding it to the \code{ggheatmap} for a collection of
numeric columns. It works if \code{track_columns} are included in \code{ggData(gghm)}.
}
