% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tracks.R
\name{add_tracks}
\alias{add_tracks}
\title{Add column tracks to a \code{ggheatmap}}
\usage{
add_tracks(
  gghm,
  track_columns,
  track_colors = list(),
  track_prop = 0.3,
  leg_ncol = 3,
  fontsize = 11,
  track_pos = "bottom",
  legend_action = "collect",
  show_legend = TRUE
)
}
\arguments{
\item{gghm}{An ggplot object of class \code{ggheatmap}.}

\item{track_columns}{A vector with names of columns to be plotted.}

\item{track_colors}{A named list where names are the same as \code{track_columns}.
Can be either:
\itemize{
\item NULL, in which case all colors will be chosen automatically.
\item A valid palette used by \code{RColorBrewer}. See: \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::display.brewer.all()}}
\item A named vector of colors passed to \code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual()}}.
}}

\item{track_prop}{A number between 0 and 1, representing the height
proportion between new tracks and the heatmap.}

\item{leg_ncol}{Number of columns in the track legends. Passed to
\code{\link[ggplot2:guide_legend]{ggplot2::guide_legend()}}.}

\item{fontsize}{Base fontsize for plot, which will be used by the theme.
Ultimately passed to \code{\link[ggplot2:ggtheme]{ggplot2::theme_minimal()}} as \code{base_size}.}

\item{track_pos}{One of: 'bottom' or 'top'.}

\item{legend_action}{A string specifying how guides should be treated in the layout.
See: guides in \code{\link[patchwork:plot_layout]{patchwork::plot_layout()}}.}

\item{If}{FALSE, no legend is added to the ggheatmap for the tracks}
}
\description{
add_tracks() is a shortcut for building a new panel of \code{ggplot2::geom_tile()}s
with \code{track_columns} and adding it to the \code{ggheatmap}. It works if \code{track_columns}
are included in \code{ggData(gghm)}.
}
