% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_heatmaps.R
\name{align_heatmaps}
\alias{align_heatmaps}
\title{Align plot to ggheatmap}
\usage{
align_heatmaps(
  hm1,
  hm2,
  pos = "horizontal",
  legend_action = NULL,
  tag_level = "new"
)
}
\arguments{
\item{hm1}{first heatmap}

\item{hm2}{second heatmap
(if pos='top' or pos='bottom') or rows (if pos='left') with the heatmap.
Note: to create the gplot, make sure you use \code{\link[=gghmData]{gghmData()}} to get the data, or
alternatively \code{\link[=get_rowLevels]{get_rowLevels()}} or \code{\link[=get_colLevels]{get_colLevels()}}.}

\item{pos}{'vertical' or 'horizontal', by default the best option.
of the height (if pos='top' or pos='bottom') or width (if pos='left') that
the new plot will occupy in the panel.}

\item{legend_action}{A string specifying how guides should be treated in the layout.
See: guides in \code{\link[patchwork:plot_layout]{patchwork::plot_layout()}}.}

\item{tag_level}{A string ('keep' or 'new') to indicate how auto-tagging should behave.
See \link[patchwork:plot_annotation]{patchwork::plot_annotation}.}
}
\value{
A joined panel object of class ggheatmap.
}
\description{
Align a ggheatmap object to another ggplot that shares either its columns or
rows to form complex panels. This function is called internally by add_tracks,
and also exported by ggheatmap.
}
