% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_to_hm.R
\name{align_to_hm}
\alias{align_to_hm}
\title{Align plot to ggheatmap}
\usage{
align_to_hm(
  gghm,
  gplot,
  pos = "bottom",
  newplt_size_prop = 0.5,
  legend_action = NULL,
  tag_level = "new"
)
}
\arguments{
\item{gghm}{A object with class ggheatmap, generated by \code{\link[=ggheatmap]{ggheatmap()}}.}

\item{gplot}{Another object of class ggplot, that shares either columns
(if \code{pos='top'} or \code{pos='bottom'}) or rows (if \code{pos='left'}) with the heatmap.
Note: to create the gplot, make sure you use \code{\link[=gghmData]{gghmData()}} to get the data, or
alternatively \code{\link[=gghm_rowLevels]{gghm_rowLevels()}} or \code{\link[=gghm_colLevels]{gghm_colLevels()}}.}

\item{pos}{One of 'top', 'bottom' or 'left'.}

\item{newplt_size_prop}{A float between 0 and 1, that indicates the proportion
of the height (if \code{pos='top'} or \code{pos='bottom'}) or width (if \code{pos='left'}) that
the new plot will occupy in the panel.}

\item{legend_action}{A string specifying how guides should be treated in the layout.
See: guides in \code{\link[patchwork:plot_layout]{patchwork::plot_layout()}}.}

\item{tag_level}{A string ('keep' or 'new') to indicate how auto-tagging should behave.
See \link[patchwork:plot_annotation]{patchwork::plot_annotation}.}
}
\value{
A joined panel object of class \code{ggheatmap}.
}
\description{
Align a ggheatmap object to another ggplot that shares either its columns or
rows to form complex panels. This function is called internally by add_tracks,
and also exported by ggheatmap.
}
