% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_extras.R
\name{annotate_gghm}
\alias{annotate_gghm}
\title{Annotate rows on gghm}
\usage{
annotate_gghm(
  gghm,
  annot_labels,
  side = "left",
  annot_size = 2.5,
  add_margin = 0
)
}
\arguments{
\item{gghm}{An ggplot object of class ggheatmap.}

\item{annot_labels}{A vector of labels that you'd like to annotate. Must have
been passed as \code{rowv} when creating the heatmap.}

\item{side}{Either 'left' or 'right' of the heatmap}

\item{annot_size}{Size of annotation text}

\item{add_margin}{Whether to add some margin to make extra space for the
annotations if they're overlapping other elements. Margin unit is mm.}
}
\description{
Adds row labels on the left or right side of the gghm. Especially useful
when there are too many \code{rowv}, but you'd still like to highlight certain
variables in the heatmap.
}
