% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data}
\alias{get_data}
\alias{gghmData}
\alias{get_rowLevels}
\alias{get_colLevels}
\title{Get data from ggheatmap}
\usage{
get_data(gghm)

gghmData(gghm)

get_rowLevels(gghm)

get_colLevels(gghm)
}
\arguments{
\item{gghm}{An object of class \code{ggheatmap}}
}
\description{
These three functions help get data that may be useful for further plotting
from the ggheatmap object.
}
\details{
\code{get_data()} recovers the table with an ID column named \code{observations}, which
if a factor that uses the same ordered levels as the heatmap. Those levels are
also provided by calling \code{get_colLevels()}. For plots where you'd like to
align the rows to the rows of the heatmap, \code{get_rowLevels()} provides the
order of the rows in the heatmap.
}
