% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gghm_getHMplot}
\alias{gghm_getHMplot}
\alias{gghm_updateHMplot}
\title{Get and update heatmap from ggheatmap}
\usage{
gghm_getHMplot(gghm)

gghm_updateHMplot(gghm, new_hm)
}
\arguments{
\item{gghm}{an object of class ggheatmap}

\item{new_hm}{an object of class ggplot, which corresponds to the heatmap
portion of the ggheatmap to overwrite the current heatmap portion.}
}
\description{
Functions to get and modify heatmap data
}
\details{
Allows the user to update graphical parameters of the heatmap by adding onto
it with ggplot syntax as they like. You first perform a \code{new_hm <- get_gghm(gghm)}
to get the object, which can then be modified and re-written by setting
\code{gghm <- update_gghm(gghm, new_hm)}
}
