% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{tcgaBLCA_ex}
\alias{tcgaBLCA_ex}
\title{A list with data extracted from the TCGA-BLCA cohort of patients with
muscle-invasive bladder cancer}
\format{
A list comprised of:
\enumerate{
\item \code{gexp}: A matrix of gene expression of 60 genes for 100 patients
\item \code{sample_annot}: A data.frame with clinical and molecular annotations
\item \code{gene_annotation}: A data.frame with annotation for the 60 genes
}
}
\source{
10.1016/j.cell.2017.09.007
}
\usage{
tcgaBLCA_ex
}
\description{
A list with data extracted from the TCGA-BLCA cohort of patients with
muscle-invasive bladder cancer
}
\keyword{datasets}
