% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_extras.R
\name{theme_sparse}
\alias{theme_sparse}
\alias{theme_sparse2}
\alias{theme_quant}
\title{ggheatmap additional themes}
\usage{
theme_sparse(...)

theme_sparse2(...)

theme_quant(...)
}
\arguments{
\item{...}{all arguments passed to \code{ggplot2::theme_minimal()}.}
}
\description{
These themes are provided as helpers to build panels with \code{ggheatmap}. They
are all built on top of \code{\link[ggplot2:ggtheme]{ggplot2::theme_minimal()}}.
}
\details{
\code{theme_sparse()} is the baseline simple and spartan theme to build plots that align
with columns of \code{ggheatmap}s. \code{theme_sparse2()} is designed for plots that will
align to rows of the \code{ggheatmap}. \code{theme_quant()} adds horizontal lines to
\code{theme_sparse()} and can be a useful upgrade for displaying numeric values in
the y-axis.
}
