% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheatmap.R
\name{ggheatmap}
\alias{ggheatmap}
\title{Plots a ggplot heatmap}
\usage{
ggheatmap(
  table,
  colv = NULL,
  rowv = NULL,
  hm_colors = "RdYlBu",
  hm_color_values = NULL,
  hm_color_breaks = waiver(),
  scale = FALSE,
  center = FALSE,
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  dist_method = "euclidean",
  clustering_method = "complete",
  raster = FALSE,
  rows_title = "",
  column_title = "",
  colors_title = "value",
  fontsize = 11,
  show_rownames = TRUE,
  show_colnames = TRUE,
  show_dend_row = TRUE,
  show_dend_col = TRUE,
  dend_lwd = 0.3,
  dend_prop_row = 0.1,
  dend_prop_col = 0.1,
  group_track = TRUE,
  group_prop = 0.1,
  group_colors = NULL,
  group_lines = FALSE,
  group_line_color = "black",
  group_lty = "solid",
  group_lwd = 0.3,
  group_leg_ncol = 3,
  row_facetting_space = 3
)
}
\arguments{
\item{table}{A table that can follow four formats:
\itemize{
\item If \code{table} is a \code{data.frame} or \code{tbl} and \code{colv} and \code{rowv} are provided,
\code{colv} will be plotted into the columns (observations), while the variables
provided as \code{rowv} will be plotted into the rows (variables).
\item If \code{table} is a \code{grouped_tbl} and \code{cluster_cols = TRUE}, a semi-supervised
clustering of the data will be performed instead only clustering within groups.
\item If \code{table} is a \code{data.frame} or \code{tbl} and only \code{colv}, all other columns
will be used as \code{rowv}.
\item If \code{table} is a \code{matrix} or \code{data.frame} and \code{colv} and \code{rowv} are \code{NULL},
\code{ggheatmap} will plot it as-is (columns will be plotted as columns, rows as
rows of the heatmap). It is convenient, but less powerful.
}}

\item{colv}{Either NULL or the name of the column that contains the
observation ID variable for the data.}

\item{rowv}{Either NULL, a vector with names of columns to be plotted into the rows
of the heatmap, or a named list of vectors for a faceted plot (\code{show_dend_row} will then
be ignored).}

\item{hm_colors}{Can be either:
\itemize{
\item A valid palette used by \code{RColorBrewer}. See: \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::display.brewer.all()}}
\item A vector of colors to be interpolated
}}

\item{hm_color_values}{Either NULL or a vector of values used for interpolation
between colors. Passed to \code{values} from \code{\link[ggplot2:scale_gradient]{ggplot2::scale_fill_gradientn()}}}

\item{hm_color_breaks}{Breaks for heatmap colors. Passed to \code{breaks} from
\code{\link[ggplot2:scale_gradient]{ggplot2::scale_fill_gradientn()}}}

\item{scale}{If TRUE, data will be scaled (z-score)}

\item{center}{If TRUE, data will be centered}

\item{cluster_rows}{If TRUE, rows will be clustered with \code{\link[stats:hclust]{stats::hclust()}}}

\item{cluster_cols}{If TRUE, columns will be clustered with \code{\link[stats:hclust]{stats::hclust()}}}

\item{dist_method}{Distance method passed to \code{\link[stats:dist]{stats::dist()}}. Also supports
correlation distance (\code{1-cor()}), with \code{dist_method} passed to method from
\code{stats::cor}}

\item{clustering_method}{Clustering method is passed to method from
\code{stats::hclust()}}

\item{raster}{If TRUE, \code{\link[ggplot2:geom_tile]{ggplot2::geom_raster()}} will be used for the
heatmap tiles instead of \code{\link[ggplot2:geom_tile]{ggplot2::geom_tile()}}. Will be recommended by
the function for large tables.}

\item{rows_title}{Label for row variables.}

\item{column_title}{Label for the column variable.}

\item{colors_title}{Labels for the color legend.}

\item{fontsize}{Base fontsize for plot, which will be used by the theme.
Ultimately passed to \code{\link[ggplot2:ggtheme]{ggplot2::theme_minimal()}} as \code{base_size}.}

\item{show_rownames}{If TRUE, row names will be shown in the heatmap.}

\item{show_colnames}{If TRUE, col names will be shown in the heatmap.}

\item{show_dend_row}{If TRUE, the clustering dendrogram for the row variables
will be shown to the left of the heatmap}

\item{show_dend_col}{If TRUE, the clustering dendrogram for the column variable
will be shown on top of the heatmap.}

\item{dend_lwd}{Linewidth for the dendrogram drawing. Passed to
\link[ggtree:ggtree]{ggtree::ggtree} as size.}

\item{dend_prop_row}{A value between 0 and 1. The proportion of the width
occupied by the dendrogram.}

\item{dend_prop_col}{A value between 0 and 1. The proportion of the height
occupied by the dendrogram.}

\item{group_track}{If \code{table} is a \code{grouped_tbl} and \code{group_track = TRUE},
a track will be plot between the dendrogram and heatmap.}

\item{group_prop}{The proportion of the height of the heatmap that will be
used for the group track.}

\item{group_colors}{A named vector with colors for each level in the grouping
variable. If NULL, automatic colors will be used.}

\item{group_leg_ncol}{Number of columns in the groups legend. Passed to
\code{\link[ggplot2:guide_legend]{ggplot2::guide_legend()}}.}
}
\value{
A ggplot object with class \code{ggheatmap}.
}
\description{
\code{ggheatmap()} is the main function of the \code{ggheatmap} package. It constructs
a tile-able plot using ggplot and patchwork, that can be added to and used in
constructing panels.
}
